git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@11738 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ io/src/writer/Hdf5DataWriter.cpp
@@ -963,7 +963,7 @@ void Hdf5DataWriter::AdvanceAlongUnlimitedDimension()
     }
 
     // If you are beyond the user estimate increment step by step
-    if ( mCurrentTimeStep >= mEstimatedUnlimitedLength )
+    if ( mCurrentTimeStep >= (long) mEstimatedUnlimitedLength )
     {
         mDatasetDims[0]++;
         mNeedExtend = true;
